Finished first tests and improved MinHeap.DecreaseKey()

+++ Algorithms/headers/BellmanFord.h
@@ -6,10 +6,10 @@ class BellmanFord
 {
   private:
 	int m_Size = 0;
-	int *m_pathWeightArr = nullptr;
+	float *m_pathWeightArr = nullptr;
 	
 	BellmanFord(int i_Size, int i_StartIndex);
-	void relax(int i_U, int i_V, int i_Weight);
+	void relax(int i_U, int i_V, float i_Weight);
 	bool isThereImprovingEdge(const Graph &i_Graph);
   
   public:
@@ -17,5 +17,5 @@ class BellmanFord
 	BellmanFord(const BellmanFord &org) = delete;
 	~BellmanFord();
 	const BellmanFord &operator=(const BellmanFord &org) = delete;
-	static int Execute(const Graph &i_Graph, int i_Start, int i_Target);
+	static float Execute(const Graph &i_Graph, int i_Start, int i_Target);
 };
