+++ DataStructures/cpp/AdjacentList.cpp
@@ -4,7 +4,8 @@ AdjacentList::~AdjacentList()
 {
 	this->freeList();
 }
-ListNode *AdjacentList::makeNewNode(int i_Vertex, int i_Weight, ListNode *io_Next)
+
+ListNode *AdjacentList::makeNewNode(int i_Vertex, float i_Weight, ListNode *io_Next)
 {
 	ListNode *newNode = new ListNode;
 	
@@ -15,7 +16,8 @@ ListNode *AdjacentList::makeNewNode(int i_Vertex, int i_Weight, ListNode *io_Nex
 	
 	return newNode;
 }
-void AdjacentList::ToHead(int i_Vertex, int i_Weight)
+
+void AdjacentList::ToHead(int i_Vertex, float i_Weight)
 {
 	ListNode *newHead = this->makeNewNode(i_Vertex, i_Weight, m_Head);
 	
@@ -31,7 +33,8 @@ void AdjacentList::ToHead(int i_Vertex, int i_Weight)
 	
 	m_Size++;
 }
-void AdjacentList::ToTail(int i_Vertex, int i_Weight)
+
+void AdjacentList::ToTail(int i_Vertex, float i_Weight)
 {
 	ListNode *newTail = this->makeNewNode(i_Vertex, i_Weight);
