+++ DataStructures/header/WeightedGraphByLists.h
@@ -22,8 +22,9 @@ namespace GraphByLists
 	  static WeightedGraph MakeEmptyGraph(int i_Size);
 	  virtual bool IsAdjacent(int i_U, int i_V) const;
 	  virtual AdjacentList GetAdjList(int i_U) const;
-	  virtual void AddEdge(int i_U, int i_V, int i_Weight);
+	  virtual void AddEdge(int i_U, int i_V, float i_Weight);
 	  virtual void RemoveEdge(int i_U, int i_V);
+	
 	  virtual int GetSize() const{ return m_Size; }
   };
 }
