+++ DataStructures/header/WeightedGraphByMat.h
@@ -8,7 +8,8 @@ namespace GraphByMat
   {
 	private:
 	  int m_Size;
-	  int **m_AdjMat;
+	  float **m_AdjMat;
+	
 	  bool isValidVertex(int i_U) const{ return i_U > 0 && i_U <= m_Size; }
   
 	public:
@@ -22,8 +23,9 @@ namespace GraphByMat
 	  static WeightedGraph MakeEmptyGraph(int i_Size);
 	  virtual bool IsAdjacent(int i_U, int i_V) const;
 	  virtual AdjacentList GetAdjList(int i_U) const;
-	  virtual void AddEdge(int i_U, int i_V, int i_Weight = MAX_WEIGHT);
+	  virtual void AddEdge(int i_U, int i_V, float i_Weight = MAX_WEIGHT);
 	  virtual void RemoveEdge(int i_U, int i_V);
+	
 	  virtual int GetSize() const{ return m_Size; }
   };
 }
